Commit Before Pagination

+++ backend/endpoints/expenditure.py
@@ -39,10 +39,10 @@ async def get_income_details(user_id:int):
     conn = None
     try:
         conn = get_connection()
-        query = f"select * from {EXPENDITURE_TABLE}"
+        query = f"select * from {EXPENDITURE_TABLE} where user_id = ? order by exp_date desc"
         cursor = conn.cursor()
 
-        res = cursor.execute(query).fetchall()
+        res = cursor.execute(query, [user_id]).fetchall()
         
 
         result = []
@@ -69,16 +69,23 @@ async def get_income_details(user_id:int):
 async def delete_entry(request:Request):
     body =await request.json()
     id = body.get("id")
-    print(id)
+    user_id = body.get("user_id")
+
+    if not id or not user_id:
+        return {"code":400,"status":"error","message":"Missing delete id or user_id"}
+
     conn = None
     try:
         conn = get_connection()
         cursor = conn.cursor()
-        query = f"delete from {EXPENDITURE_TABLE} where id = ?"
-        res = cursor.execute(query,[id])
+        query = f"delete from {EXPENDITURE_TABLE} where id = ? and user_id = ?"
+        res = cursor.execute(query,[id, user_id])
         conn.commit()
+
+        if cursor.rowcount == 0:
+            return {"code":404,"status":"error","message":"Entry not found or unauthorized"}
     except Exception as e:
-        return {"code":400,"status":"error"}
+        return {"code":400,"status":"error","message":"Error deleting entry"}
     finally:
         if conn:
             if cursor:
